+++ NEWS.md
@@ -30,6 +30,7 @@
 * Fixed issue with autocompletion of column names within native-piped R expressions (#9385)
 * Fixed issue where help requests for Python objects would fail with reticulate 1.20 (#9311)
 * Fixed issue where busy sessions can't be interrupted and block basic file operations (#2038)
+* Fixed issue where R Markdown error output was not properly formatted when displayed (#9390)
 * Fixed issue where R Markdown template skeletons with a '.rmd' extension were not discovered (Pro #1607)
 * Fixed issues causing multiple background jobs to be created when running Shiny applications in the background (#8746, #6904)
 * Fixed issue causing an error when adding files to static content published to RStudio Connect (#9571)